refactor : SeriesDetailsScreen separate it into component
