+++ presentation/src/main/java/com/madrid/presentation/screens/detailsScreen/seriesDetails/component/CopyToClipboard.kt
@@ -1,0 +1,13 @@
+package com.madrid.presentation.screens.detailsScreen.seriesDetails.component
+
+import android.content.ClipData
+import android.content.ClipboardManager
+import android.content.Context
+import android.widget.Toast
+
+fun copyToClipboard(text: String, context: Context) {
+    val clipboard = context.getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
+    val clip = ClipData.newPlainText("Series Link", text)
+    clipboard.setPrimaryClip(clip)
+    Toast.makeText(context, "Link copied", Toast.LENGTH_SHORT).show()
+}
