+++ presentation/src/main/java/com/madrid/presentation/screens/detailsScreen/seriesDetails/component/DescriptionSection.kt
@@ -1,0 +1,21 @@
+package com.madrid.presentation.screens.detailsScreen.seriesDetails.component
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.madrid.designSystem.component.TextWithReadMore
+import com.madrid.presentation.viewModel.detailsViewModel.SeriesDetails.SeriesDetailsUiState
+
+@Composable
+fun DescriptionSection(uiState: SeriesDetailsUiState) {
+    if (uiState.description.isNotEmpty()) {
+        TextWithReadMore(
+            description = uiState.description,
+            modifier = Modifier
+                .padding(horizontal = 16.dp)
+                .padding(bottom = 32.dp),
+            maxLines = 5
+        )
+    }
+}
