+++ presentation/src/main/java/com/madrid/presentation/screens/detailsScreen/seriesDetails/component/ShareToApp.kt
@@ -1,0 +1,21 @@
+package com.madrid.presentation.screens.detailsScreen.seriesDetails.component
+
+import android.content.Context
+import android.content.Intent
+
+fun shareToApp(appPackage: String, url: String, context: Context) {
+    val intent = Intent(Intent.ACTION_SEND).apply {
+        type = "text/plain"
+        putExtra(Intent.EXTRA_TEXT, url)
+        setPackage(appPackage)
+    }
+    try {
+        context.startActivity(intent)
+    } catch (e: Exception) {
+        val fallback = Intent(Intent.ACTION_SEND).apply {
+            type = "text/plain"
+            putExtra(Intent.EXTRA_TEXT, url)
+        }
+        context.startActivity(Intent.createChooser(fallback, "Share via"))
+    }
+}
